LPD-103394 Stop a large cluster node count from overflowing the quota check
A license key's cluster node count arrives in the request and nothing
bounded it. The quota check counts a clustered key as its node count and
summed those into an int, so a request asking for a node count near the
maximum wrapped the total negative, the comparison against the
entitlement's quantity read as satisfied, and the check waved through the
very key it exists to stop.

The count is now capped and the total accumulates in a long, so neither
the cap nor the arithmetic depends on the other holding.

+++ workspaces/liferay-one-workspace/client-extensions/liferay-one-etc-spring-boot/src/main/java/com/liferay/one/AccountsRestController.java
@@ -466,7 +466,7 @@ public class AccountsRestController extends OneBaseRestController {
 			account, "allowPermanentLicenses", true);
 		boolean complimentary = false;
 		List<Entitlement> entitlements = new ArrayList<>();
-		Map<Long, Integer> pendingServerCounts = new HashMap<>();
+		Map<Long, Long> pendingServerCounts = new HashMap<>();
 
 		for (int i = 0; i < jsonArray.length(); i++) {
 			JSONObject jsonObject = jsonArray.getJSONObject(i);
@@ -1148,7 +1148,7 @@ public class AccountsRestController extends OneBaseRestController {
 
 	private void _validateLicenseKey(
 			boolean allowPermanentLicenses, Entitlement entitlement,
-			JSONObject jsonObject, Map<Long, Integer> pendingServerCounts)
+			JSONObject jsonObject, Map<Long, Long> pendingServerCounts)
 		throws Exception {
 
 		_licenseKeyEntitlementValidator.validateTerm(
